cleanup group by examples
